fix(skore): Merge conftest.py
